requirements.txt
fpdf==1.7.2
pycups==2.0.1
requests==2.32.2